requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
arrow==1.2.1
    # via jinja2-time
binaryornot==0.4.4
    # via cookiecutter
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via
    #   cryptography
    #   pynacl
chardet==4.0.0
    # via binaryornot
charset-normalizer==2.0.7
    # via requests
click==8.0.3
    # via
    #   cookiecutter
    #   cruft
    #   typer
cookiecutter==1.7.3
    # via
    #   -r requirements.in
    #   cruft
cruft==2.10.1
    # via -r requirements.in
cryptography==36.0.1
    # via -r requirements.in
gitdb==4.0.9
    # via gitpython
gitpython==3.1.24
    # via cruft
idna==3.3
    # via requests
importlib-metadata==3.10.1
    # via cruft
jinja2==3.0.2
    # via
    #   cookiecutter
    #   jinja2-time
jinja2-time==0.2.0
    # via cookiecutter
markupsafe==2.0.1
    # via jinja2
poyo==0.5.0
    # via cookiecutter
pycparser==2.21
    # via cffi
pynacl==1.5.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via arrow
python-slugify==5.0.2
    # via cookiecutter
requests==2.26.0
    # via cookiecutter
sh==1.14.2
    # via -r requirements.in
six==1.16.0
    # via
    #   cookiecutter
    #   python-dateutil
smmap==5.0.0
    # via gitdb
text-unidecode==1.3
    # via python-slugify
typer==0.4.0
    # via cruft
typing-extensions==3.10.0.2
    # via gitpython
urllib3==1.26.7
    # via requests
zipp==3.6.0
    # via importlib-metadata
